refactor + auto OCR
