+++ ingestion/loaders/image_loader.py
@@ -1,26 +1,26 @@
 """
-Smart image loader that processes standalone images with intelligent OCR.
+Image loader that processes standalone images with intelligent OCR.
 """
 
 import logging
 from pathlib import Path
 from typing import Any, Dict, Optional
 
-from core.smart_ocr import smart_ocr_processor
+from core.ocr import ocr_processor
 
 logger = logging.getLogger(__name__)
 
 
-class SmartImageLoader:
+class ImageLoader:
     """Loads content from image files with intelligent OCR detection."""
 
     def __init__(self):
-        """Initialize the smart image loader."""
-        self.processor = smart_ocr_processor
+        """Initialize the image loader."""
+        self.processor = ocr_processor
 
     def load(self, file_path: Path) -> Optional[str]:
         """
-        Load text content from an image file using smart OCR.
+        Load text content from an image file using intelligent OCR.
 
         Args:
             file_path: Path to the image file
@@ -33,7 +33,7 @@ class SmartImageLoader:
 
     def load_with_metadata(self, file_path: Path) -> Optional[Dict[str, Any]]:
         """
-        Load image content with smart OCR and detailed metadata.
+        Load image content with intelligent OCR and detailed metadata.
 
         Args:
             file_path: Path to the image file
@@ -42,9 +42,9 @@ class SmartImageLoader:
             Dictionary with content and OCR metadata
         """
         try:
-            logger.info(f"Processing image with smart OCR: {file_path}")
+            logger.info(f"Processing image with intelligent OCR: {file_path}")
             
-            # Use smart OCR processor to analyze and extract from image
+            # Use OCR processor to analyze and extract from image
             result = self.processor.process_standalone_image(file_path)
             
             if not result["content"]:
@@ -56,7 +56,7 @@ class SmartImageLoader:
             content_analysis = ocr_metadata.get("content_analysis", {})
             
             metadata = {
-                "processing_method": "smart_image_ocr",
+                "processing_method": "image_ocr",
                 "file_type": "standalone_image",
                 "ocr_applied": ocr_metadata.get("ocr_applied", 0),
                 "ocr_items_count": len(ocr_metadata.get("ocr_items", [])),
@@ -68,7 +68,7 @@ class SmartImageLoader:
             # Log processing result
             if ocr_metadata.get("ocr_applied", 0) > 0:
                 content_type = metadata["content_primary_type"]
-                logger.info(f"Smart OCR extracted text from {content_type} image: {file_path}")
+                logger.info(f"OCR extracted text from {content_type} image: {file_path}")
             
             return {
                 "content": result["content"],
@@ -76,5 +76,5 @@ class SmartImageLoader:
             }
 
         except Exception as e:
-            logger.error(f"Failed to load image with smart OCR {file_path}: {e}")
+            logger.error(f"Failed to load image with OCR {file_path}: {e}")
             return None
